Fix some compile errors
git-svn-id: https://svn.atnf.csiro.au/difx@3809 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/mk5daemon/trunk/src/mk5daemon.h
@@ -116,7 +116,7 @@ extern const char MacListCommandStrings[][10];
 class MAC
 {
 public:
-	long int address;
+	unsigned long long int address;
 public:
 	int parse(const char *str);
 	int toString(char *str) const;	/* returns number of characters written */
